backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.2
pillow==11.1.0
Werkzeug==3.1.3
